Add all Store operations and created a file for testing
