+++ src/main.rs
@@ -1,8 +1,9 @@
 use std::env;
 use std::ops::{Index, IndexMut};
-pub mod file_management;
-pub mod operations;
 use operations::*;
+pub mod file_management;
+mod operations;
+mod tests;
 static MEM_MAX: usize = 1 << 16;
 static PC_START: u16 = 0x3000;
 #[derive(Clone, Copy)]
@@ -134,7 +135,7 @@ fn main() -> Result<(), String> {
             Operations::Jsr => jump_to_subrutine(instruction, &mut state),
             Operations::And => and(instruction, &mut state),
             Operations::Ldr => load_register(instruction, &mut state),
-            Operations::Str => todo!(), //store_register(instruction),
+            Operations::Str => store_register(instruction,&mut state),
             Operations::Rti => todo!(), 
             Operations::Not => not(instruction,&mut state),
             Operations::Ldi => load_indirect(instruction, &mut state),
